+++ mobile/android/android-components/components/browser/storage-sync/src/test/java/mozilla/components/browser/storage/sync/RemoteTabsCommandQueueTest.kt
@@ -26,7 +26,7 @@ import org.mockito.Mockito.verify
 class RemoteTabsCommandQueueTest {
     private val storage: RemoteTabsStorage = mock()
     private val api: RemoteCommandStore = mock()
-    private val closeTabsCommandSender: RemoteTabsCommandQueue.CommandSender<DeviceCommandOutgoing.CloseTab> = mock()
+    private val closeTabsCommandSender: RemoteTabsCommandQueue.CommandSender<DeviceCommandOutgoing.CloseTab, RemoteTabsCommandQueue.SendCloseTabsResult> = mock()
     private val commands = RemoteTabsCommandQueue(storage, closeTabsCommandSender)
 
     @Before
@@ -63,7 +63,7 @@ class RemoteTabsCommandQueueTest {
     }
 
     @Test
-    fun `GIVEN a queue with unsent commands AND a sender that sends all commands WHEN the queue is flushed THEN the flush should succeed`() = runTest {
+    fun `GIVEN a queue with unsent pending commands AND a sender that sends all commands WHEN the queue is flushed THEN the flush should succeed`() = runTest {
         whenever(api.getUnsentCommands()).thenReturn(
             listOf(
                 PendingCommand(
@@ -80,7 +80,7 @@ class RemoteTabsCommandQueueTest {
                 ),
             ),
         )
-        whenever(closeTabsCommandSender.send(any(), any())).thenReturn(RemoteTabsCommandQueue.SendResult.Ok)
+        whenever(closeTabsCommandSender.send(any(), any())).thenReturn(RemoteTabsCommandQueue.SendCloseTabsResult.Ok)
         whenever(api.setPendingCommandSent(any())).thenReturn(true)
 
         assertEquals(DeviceCommandQueue.FlushResult.ok(), commands.flush())
@@ -89,7 +89,7 @@ class RemoteTabsCommandQueueTest {
     }
 
     @Test
-    fun `GIVEN a queue with unsent commands AND a sender that returns 'no account' WHEN the queue is flushed THEN the flush should succeed`() = runTest {
+    fun `GIVEN a queue with unsent pending commands AND a sender that returns 'no account' WHEN the queue is flushed THEN the flush should succeed`() = runTest {
         whenever(api.getUnsentCommands()).thenReturn(
             listOf(
                 PendingCommand(
@@ -100,7 +100,7 @@ class RemoteTabsCommandQueueTest {
                 ),
             ),
         )
-        whenever(closeTabsCommandSender.send(any(), any())).thenReturn(RemoteTabsCommandQueue.SendResult.NoAccount)
+        whenever(closeTabsCommandSender.send(any(), any())).thenReturn(RemoteTabsCommandQueue.SendCloseTabsResult.NoAccount)
 
         assertEquals(DeviceCommandQueue.FlushResult.ok(), commands.flush())
         verify(closeTabsCommandSender).send(any(), any())
@@ -108,7 +108,7 @@ class RemoteTabsCommandQueueTest {
     }
 
     @Test
-    fun `GIVEN a queue with unsent commands AND a sender that returns 'no device' WHEN the queue is flushed THEN the flush should succeed`() = runTest {
+    fun `GIVEN a queue with unsent pending commands AND a sender that returns 'no device' WHEN the queue is flushed THEN the flush should succeed`() = runTest {
         whenever(api.getUnsentCommands()).thenReturn(
             listOf(
                 PendingCommand(
@@ -119,7 +119,7 @@ class RemoteTabsCommandQueueTest {
                 ),
             ),
         )
-        whenever(closeTabsCommandSender.send(any(), any())).thenReturn(RemoteTabsCommandQueue.SendResult.NoDevice)
+        whenever(closeTabsCommandSender.send(any(), any())).thenReturn(RemoteTabsCommandQueue.SendCloseTabsResult.NoDevice)
 
         assertEquals(DeviceCommandQueue.FlushResult.ok(), commands.flush())
         verify(closeTabsCommandSender).send(any(), any())
@@ -127,7 +127,7 @@ class RemoteTabsCommandQueueTest {
     }
 
     @Test
-    fun `GIVEN a queue with unsent commands AND a sender that returns an error WHEN the queue is flushed THEN the flush should be retried`() = runTest {
+    fun `GIVEN a queue with unsent pending commands AND a sender that returns an error WHEN the queue is flushed THEN the flush should be retried`() = runTest {
         whenever(api.getUnsentCommands()).thenReturn(
             listOf(
                 PendingCommand(
@@ -138,7 +138,7 @@ class RemoteTabsCommandQueueTest {
                 ),
             ),
         )
-        whenever(closeTabsCommandSender.send(any(), any())).thenReturn(RemoteTabsCommandQueue.SendResult.Error)
+        whenever(closeTabsCommandSender.send(any(), any())).thenReturn(RemoteTabsCommandQueue.SendCloseTabsResult.Error)
 
         assertEquals(DeviceCommandQueue.FlushResult.retry(), commands.flush())
         verify(closeTabsCommandSender).send(any(), any())
@@ -146,7 +146,7 @@ class RemoteTabsCommandQueueTest {
     }
 
     @Test
-    fun `GIVEN a queue with unsent commands AND a sender that sends some commands WHEN the queue is flushed THEN the flush should be retried`() = runTest {
+    fun `GIVEN a queue with unsent pending commands AND a sender that sends some commands WHEN the queue is flushed THEN the flush should be retried`() = runTest {
         whenever(api.getUnsentCommands()).thenReturn(
             listOf(
                 PendingCommand(
@@ -163,8 +163,8 @@ class RemoteTabsCommandQueueTest {
                 ),
             ),
         )
-        whenever(closeTabsCommandSender.send(eq("123"), any())).thenReturn(RemoteTabsCommandQueue.SendResult.Ok)
-        whenever(closeTabsCommandSender.send(eq("1234"), any())).thenReturn(RemoteTabsCommandQueue.SendResult.Error)
+        whenever(closeTabsCommandSender.send(eq("123"), any())).thenReturn(RemoteTabsCommandQueue.SendCloseTabsResult.Ok)
+        whenever(closeTabsCommandSender.send(eq("1234"), any())).thenReturn(RemoteTabsCommandQueue.SendCloseTabsResult.Error)
         whenever(api.setPendingCommandSent(any())).thenReturn(true)
 
         assertEquals(DeviceCommandQueue.FlushResult.retry(), commands.flush())
@@ -172,5 +172,77 @@ class RemoteTabsCommandQueueTest {
         val pendingCommandCaptor = argumentCaptor<PendingCommand>()
         verify(api).setPendingCommandSent(pendingCommandCaptor.capture())
         assertEquals("123", pendingCommandCaptor.value.deviceId)
+    }
+
+    @Test
+    fun `GIVEN a queue with unsent pending commands for the same device AND a sender that sends all commands WHEN the queue is flushed THEN the flush should succeed`() = runTest {
+        whenever(api.getUnsentCommands()).thenReturn(
+            listOf(
+                PendingCommand(
+                    deviceId = "123",
+                    command = RemoteCommand.CloseTab("https://getfirefox.com"),
+                    timeRequested = 0,
+                    timeSent = null,
+                ),
+                PendingCommand(
+                    deviceId = "123",
+                    command = RemoteCommand.CloseTab("https://getthunderbird.com"),
+                    timeRequested = 0,
+                    timeSent = null,
+                ),
+                PendingCommand(
+                    deviceId = "123",
+                    command = RemoteCommand.CloseTab("https://example.com"),
+                    timeRequested = 0,
+                    timeSent = null,
+                ),
+            ),
+        )
+
+        whenever(closeTabsCommandSender.send(eq("123"), any())).thenReturn(RemoteTabsCommandQueue.SendCloseTabsResult.Ok)
+        whenever(api.setPendingCommandSent(any())).thenReturn(true)
+
+        assertEquals(DeviceCommandQueue.FlushResult.ok(), commands.flush())
+        val outgoingCommandCaptor = argumentCaptor<DeviceCommandOutgoing.CloseTab>()
+        verify(closeTabsCommandSender).send(any(), outgoingCommandCaptor.capture())
+        verify(api, times(3)).setPendingCommandSent(any())
+
+        assertEquals(listOf("https://getfirefox.com", "https://getthunderbird.com", "https://example.com"), outgoingCommandCaptor.value.urls)
+    }
+
+    @Test
+    fun `GIVEN a queue with unsent pending commands for the same device AND a sender that returns URLs to retry WHEN the queue is flushed THEN the flush should be retried`() = runTest {
+        whenever(api.getUnsentCommands()).thenReturn(
+            listOf(
+                PendingCommand(
+                    deviceId = "123",
+                    command = RemoteCommand.CloseTab("https://getfirefox.com"),
+                    timeRequested = 0,
+                    timeSent = null,
+                ),
+                PendingCommand(
+                    deviceId = "123",
+                    command = RemoteCommand.CloseTab("https://getthunderbird.com"),
+                    timeRequested = 0,
+                    timeSent = null,
+                ),
+                PendingCommand(
+                    deviceId = "123",
+                    command = RemoteCommand.CloseTab("https://example.com"),
+                    timeRequested = 0,
+                    timeSent = null,
+                ),
+            ),
+        )
+
+        whenever(closeTabsCommandSender.send(eq("123"), any())).thenReturn(RemoteTabsCommandQueue.SendCloseTabsResult.RetryFor(listOf("https://getthunderbird.com")))
+        whenever(api.setPendingCommandSent(any())).thenReturn(true)
+
+        assertEquals(DeviceCommandQueue.FlushResult.retry(), commands.flush())
+        verify(closeTabsCommandSender).send(any(), any())
+        val pendingCommandCaptor = argumentCaptor<PendingCommand>()
+        verify(api, times(2)).setPendingCommandSent(pendingCommandCaptor.capture())
+        assertEquals(RemoteCommand.CloseTab("https://getfirefox.com"), pendingCommandCaptor.allValues[0].command)
+        assertEquals(RemoteCommand.CloseTab("https://example.com"), pendingCommandCaptor.allValues[1].command)
     }
 }