+++ mobile/android/android-components/components/feature/accounts-push/src/main/java/mozilla/components/feature/accounts/push/CloseTabsCommandReceiver.kt
@@ -6,7 +6,6 @@ package mozilla.components.feature.accounts.push
 
 import mozilla.components.browser.state.action.TabListAction
 import mozilla.components.browser.state.selector.normalTabs
-import mozilla.components.browser.state.state.TabSessionState
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.sync.DeviceCommandIncoming
 import mozilla.components.concept.sync.DeviceConstellation
@@ -30,7 +29,8 @@ class CloseTabsCommandReceiver(
      */
     fun receive(command: DeviceCommandIncoming.TabsClosed) {
         val openTabs = browserStore.state.normalTabs
-        val tabsToRemove = getTabsToRemove(openTabs, command.urls).ifEmpty { return }
+        val urlsToClose = command.urls.toSet()
+        val tabsToRemove = openTabs.filter { urlsToClose.contains(it.content.url) }.ifEmpty { return }
         val remainingTabsCount = openTabs.size - tabsToRemove.size
         val willCloseLastTab = tabsToRemove.any { it.id == browserStore.state.selectedTabId } && remainingTabsCount <= 0
 
@@ -54,19 +54,4 @@ class CloseTabsCommandReceiver(
         /** The last tab on this device was closed. */
         fun onLastTabClosed() = Unit
     }
-}
-
-private fun getTabsToRemove(openTabs: List<TabSessionState>, remotelyClosedUrls: List<String>): List<TabSessionState> {
-    // The user might have the same URL open in multiple tabs on this device, and might want
-    // to remotely close some or all of those tabs. Synced tabs don't carry enough
-    // information to know which duplicates the user meant to close, so we use a heuristic:
-    // if a URL appears N times in the remotely closed URLs list, we'll close up to
-    // N instances of that URL.
-    val countsByUrl = remotelyClosedUrls.groupingBy { it }.eachCount()
-    return openTabs
-        .filter { !it.content.private }
-        .groupBy { it.content.url }
-        .flatMap { (url, tabs) ->
-            countsByUrl[url]?.let { count -> tabs.take(count) } ?: emptyList()
-        }
 }